Update to login page
Add registration link to the login page. Link is attached to the ‘Join us’ text.

The current Plugin that we were using required us to edit the ‘wp-config.php’ file. Based on this restriction it appears we won’t be able to use this plugin to connect to the database.