Database connection
In this update, Carlos was able to update the functions.php file make the json file for the list of members.
Also, I fix the complier error when running the application. So the app runs and we are able to test the logins.  Next things to do is to make sure the logins are getting pulled correctly and logins are working on the app.